requirements.txt
Django==1.8
djangorestframework==3.1.1
django-getenv==1.3.1
psycopg2==2.6
jsonschema==2.4.0
django-cors-headers==1.0.0
django-nose==1.4
celery-with-redis==3.0
django-sslify==0.2.7
-e git+git://github.com/DariuszAniszewski/django-payu.git#egg=django-payu
